pyproject.toml
[tool.poetry]
name = "clinic-appointment-app"
version = "0.1.0"
description = "Clinic Appointment Applications"
authors = ["Alfarih Faza <alfarihfz@gmail.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python-dotenv = "^1.0.0"
python = "^3.11"
dihub-cqrs = "^0.1.2"
fastapi = "^0.103.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
psycopg = { extras = ["binary", "pool"], version = "^3.1.12" }
orjson = "^3.9.7"
python-slugify = "^8.0.1"
jwt = "^1.3.1"
dihub = "^0.1.8"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
alembic = "^1.12.0"
faker = "^19.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "main:dev"
start = "main:start"